Use a sparse array for the chapter marks. Invalidate chapter data to force a re-scan

+++ audiobook/src/main/java/de/ph1b/audiobook/features/bookPlaying/BookPlayChapter.kt
@@ -4,8 +4,8 @@ import java.io.File
 
 data class BookPlayChapter(
     val file: File,
-    val start: Long,
-    val stop: Long,
+    val start: Int,
+    val stop: Int,
     val name: String) {
 
   val duration = stop - start
